ec/dell/mec5035/acpi/ec.asl: Add mailbox interface code
Add an ACPI implementation of the mailbox interface driver in mec5035.c.
The vendor firmware only uses this during system initialization as well
as SMI handlers. To minimize the reliance of SMM, add a driver for this
interface to ACPI code so that functionality in the vendor SMI handlers
can be reimplemented in ACPI instead.

Change-Id: I29049fff0ba618d5283086c19939b534f496db9f
Signed-off-by: Nicholas Chin <nic.c3.14@gmail.com>