ec/dell/mec5035/acpi: Add Fan Device
Implement an ACPI Fan Device to report the fan speed. Vendor
firmware handles the fan in SMM and reports the fan speed through the
dell-smm-hwmon driver. Move this functionality into ACPI code to
minimize usage of SMM. Note that this does break fan speed reporting in
programs such as lm-sensors as Linux does not (yet) have a hwmon driver
for ACPI Fan devices. The fan speed can be read from the /sys/bus/acpi/*
device created for the fan.

Change-Id: Idabc6722e532be4ff03fa557e80074872cef2449
Signed-off-by: Nicholas Chin <nic.c3.14@gmail.com>