ec/dell/mec5035: Add command to disable EC-initiated thermal shutdown
If command 0xBF isn't sent, the EC unconditionally shuts down the system
as soon as the CPU temperature reaches about 87 degrees, without the CPU
throttling. With vendor firmware, the CPU is able to reach around 100
degrees before thermal throttling.

This command was found by collecting LPC bus logs while booting with
vendor firmware and then systematically replaying them with coreboot
until the command to disable the shutdown was isolated. The exact
meaning of the parameters for this command are unknown at this time, but
do seem to differ between different generations of these laptops. Due to
this, the commmand should be called by mainboard specific code which
passes the specific parameter value used.

The Google Wilco EC code, which runs on Latitude Chromebooks and seems
to be similar to the standard Latitude ECs, suggests command 0xBF tells
the EC about the CPUID, but the values observed in LPC bus logs don't
seem to match any CPUID values on the non-Chromebook systems I tested.

Observed command parameter values (sent on mailbox registers 2-4):
- E6430 (Ivy Bridge): 0x07, 0x00, 0x00
- M6800 (Haswell): 0x14, 0x00, 0x00

Change-Id: I42f09a3ef681007f64d9c5b1a29248b594737a86
Signed-off-by: Nicholas Chin <nic.c3.14@gmail.com>